Refactor: Löschen von unnötigen Sachen
