+++ src/programm/Main.java
@@ -13,7 +13,6 @@ public class Main {
         Darsteller darsteller = new Darsteller(spielleiter);
         Grundbuch grundbuch = new Grundbuch();
         Kartenmanager kartenmanager = new Kartenmanager();
-        Organisator organisator = new Organisator(spielleiter, darsteller, würfel, grundbuch, kartenmanager);
         Organisator2 organisator2 = new Organisator2(spielleiter, darsteller, würfel, grundbuch, kartenmanager);
 
         // eigentliches Spiel starten
