+++ src/programm/system/Darsteller.java
@@ -1,6 +1,5 @@
 package programm.system;
 
-import programm.system.enums.Felder;
 import programm.system.interfaces.IDarsteller;
 
 import java.util.LinkedList;
