--- src/programm/system/enums/Feldtyp.java
@@ -1,7 +1,0 @@
-package programm.system.enums;
-
-public enum Feldtyp {
-    mussFeld,
-    kaufFeld,
-    freiesFeld
-}
