+++ src/programm/system/interfaces/IGrundstück.java
@@ -1,6 +1,6 @@
 package programm.system.interfaces;
 
-import programm.system.enums.Felder;
+import programm.system.Felder;
 
 public interface IGrundstück {
     String getName();
